Suppress compiler warnings
fileio.cc: In function ‘package bf_file_read(Var, Byte, void*, Objid)’:
fileio.cc:772:47: warning: comparison between signed and unsigned integer expressions [-Wsign-compare]
   read_length = (record_length > sizeof(buffer)) ? sizeof(buffer) : record_length;

+++ fileio.cc
@@ -769,7 +769,7 @@ bf_file_read(Var arglist, Byte next, void *vdata, Objid progr)
 
   errno = 0;
 
-  read_length = (record_length > sizeof(buffer)) ? sizeof(buffer) : record_length;
+  read_length = (record_length > (int)sizeof(buffer)) ? sizeof(buffer) : record_length;
 
   if(str == 0)
 	 str = new_stream(FILE_IO_BUFFER_LENGTH);
